+++ packages/components/src/components/Button/stories/Default.stories.tsx
@@ -1,8 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import Button from "../Button";
-import { faStar } from "@fortawesome/free-regular-svg-icons/faStar";
 import React from "react";
-import { Icon } from "@/components/Icon";
+import { IconPlus } from "@/components/Icon/components/icons";
 import { action } from "@storybook/addon-actions";
 
 const meta: Meta<typeof Button> = {
@@ -36,22 +35,22 @@ export const Default: Story = {};
 
 export const Small: Story = {
   args: {
-    size: "small",
+    size: "s",
   },
 };
 
 export const WithIcon: Story = {
   render: (props) => (
     <Button {...props} aria-label="Add to favorites">
-      <Icon faIcon={faStar} />
+      <IconPlus />
     </Button>
   ),
 };
 
 export const SmallWithIcon: Story = {
   render: (props) => (
-    <Button {...props} aria-label="Add to favorites" size="small">
-      <Icon faIcon={faStar} />
+    <Button {...props} aria-label="Add to favorites" size="s">
+      <IconPlus />
     </Button>
   ),
 };