+++ packages/components/src/components/Icon/components/icons/IconProject.tsx
@@ -1,0 +1,15 @@
+/* eslint-disable */
+/* auto-generated file */
+import React, { ComponentProps, FC } from "react";
+import { IconArchive as Tabler } from "@tabler/icons-react";
+import { Icon } from "@/components/Icon";
+
+export const IconProject: FC<Omit<ComponentProps<typeof Icon>, "children">> = (
+  props,
+) => (
+  <Icon {...props}>
+    <Tabler />
+  </Icon>
+);
+
+export default IconProject;