+++ packages/components/src/components/Tooltip/stories/Default.stories.tsx
@@ -2,8 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 import Tooltip, { TooltipTrigger } from "../index";
 import React from "react";
 import { Button } from "@/components/Button";
-import { Icon } from "@/components/Icon";
-import { faSave } from "@fortawesome/free-regular-svg-icons/faSave";
+import { IconCopy } from "@/components/Icon/components/icons";
 
 const meta: Meta<typeof Tooltip> = {
   title: "Overlays/Tooltip",
@@ -13,8 +12,8 @@ const meta: Meta<typeof Tooltip> = {
   },
   render: (props) => (
     <TooltipTrigger {...props}>
-      <Button aria-label="save">
-        <Icon faIcon={faSave} />
+      <Button aria-label="copy">
+        <IconCopy />
       </Button>
       <Tooltip>Save</Tooltip>
     </TooltipTrigger>