+++ packages/components/src/components/Tooltip/stories/EdgeCases.stories.tsx
@@ -2,10 +2,9 @@ import type { Meta, StoryObj } from "@storybook/react";
 import Tooltip, { TooltipTrigger } from "../index";
 import React from "react";
 import { Button } from "@/components/Button";
-import { Icon } from "@/components/Icon";
+import { IconCopy } from "@/components/Icon/components/icons";
 import defaultMeta from "./Default.stories";
 import { dummyText } from "@/lib/dev/dummyText";
-import { faSave } from "@fortawesome/free-regular-svg-icons/faSave";
 
 const meta: Meta<typeof Tooltip> = {
   ...defaultMeta,
@@ -15,8 +14,8 @@ const meta: Meta<typeof Tooltip> = {
   },
   render: (props) => (
     <TooltipTrigger {...props}>
-      <Button aria-label="save">
-        <Icon faIcon={faSave} />
+      <Button aria-label="copy">
+        <IconCopy />
       </Button>
       <Tooltip>{dummyText.medium}</Tooltip>
     </TooltipTrigger>