+++ packages/docs/src/content/02-components/icon/examples/font-awesome.tsx
@@ -1,4 +1,0 @@
-import Icon from "@mittwald/flow-react-components/Icon";
-import faStar from "@fortawesome/free-regular-svg-icons/faStar";
-
-<Icon faIcon={faStar} />;