+++ packages/docs/src/lib/liveCode/dynamicImports.ts
@@ -2,165 +2,44 @@
 /* auto-generated file */
 import { ImportMapping } from "@/lib/liveCode/types";
 import { lazy } from "react";
-import { faCirclePlay as I7 } from "@fortawesome/free-solid-svg-icons/faCirclePlay";
-import { faStar as I9 } from "@fortawesome/free-regular-svg-icons/faStar";
-import { faSnowboarding as I13 } from "@fortawesome/free-solid-svg-icons/faSnowboarding";
-import { faSnowplow as I14 } from "@fortawesome/free-solid-svg-icons/faSnowplow";
-import { NavigationItem as I22 } from "@mittwald/flow-react-components/Navigation";
-import { faUser as I23 } from "@fortawesome/free-regular-svg-icons/faUser";
-import { faHardDrive as I24 } from "@fortawesome/free-regular-svg-icons/faHardDrive";
-import { faSave as I35 } from "@fortawesome/free-regular-svg-icons/faSave";
+import { NavigationItem as I21 } from "@mittwald/flow-react-components/Navigation";
 
 export const liveCodeEditorGlobalImports: ImportMapping = {
-  "ExampleSvg:@/content/01-getting-started/stylesheet/examples/components/ExampleSvg":
-    lazy(() =>
-      import(
-        "@/content/01-getting-started/stylesheet/examples/components/ExampleSvg"
-        ).then((module) => ({ default: module.ExampleSvg })),
-    ),
-  "Avatar:@mittwald/flow-react-components/Avatar": lazy(() =>
-    import("@mittwald/flow-react-components/Avatar").then((module) => ({
-      default: module.Avatar,
-    })),
-  ),
-  "Initials:@mittwald/flow-react-components/Initials": lazy(() =>
-    import("@mittwald/flow-react-components/Initials").then((module) => ({
-      default: module.Initials,
-    })),
-  ),
-  "Image:@mittwald/flow-react-components/Image": lazy(() =>
-    import("@mittwald/flow-react-components/Image").then((module) => ({
-      default: module.Image,
-    })),
-  ),
-  "Badge:@mittwald/flow-react-components/Badge": lazy(() =>
-    import("@mittwald/flow-react-components/Badge").then((module) => ({
-      default: module.Badge,
-    })),
-  ),
-  "Text:@mittwald/flow-react-components/Text": lazy(() =>
-    import("@mittwald/flow-react-components/Text").then((module) => ({
-      default: module.Text,
-    })),
-  ),
-  "Icon:@mittwald/flow-react-components/Icon": lazy(() =>
-    import("@mittwald/flow-react-components/Icon").then((module) => ({
-      default: module.Icon,
-    })),
-  ),
-  "faCirclePlay:@fortawesome/free-solid-svg-icons/faCirclePlay": I7,
-  "Button:@mittwald/flow-react-components/Button": lazy(() =>
-    import("@mittwald/flow-react-components/Button").then((module) => ({
-      default: module.Button,
-    })),
-  ),
-  "faStar:@fortawesome/free-regular-svg-icons/faStar": I9,
-  "Checkbox:@mittwald/flow-react-components/Checkbox": lazy(() =>
-    import("@mittwald/flow-react-components/Checkbox").then((module) => ({
-      default: module.Checkbox,
-    })),
-  ),
-  "LayoutCard:@mittwald/flow-react-components/LayoutCard": lazy(() =>
-    import("@mittwald/flow-react-components/LayoutCard").then((module) => ({
-      default: module.LayoutCard,
-    })),
-  ),
-  "Heading:@mittwald/flow-react-components/Heading": lazy(() =>
-    import("@mittwald/flow-react-components/Heading").then((module) => ({
-      default: module.Heading,
-    })),
-  ),
-  "faSnowboarding:@fortawesome/free-solid-svg-icons/faSnowboarding": I13,
-  "faSnowplow:@fortawesome/free-solid-svg-icons/faSnowplow": I14,
-  "InlineAlert:@mittwald/flow-react-components/InlineAlert": lazy(() =>
-    import("@mittwald/flow-react-components/InlineAlert").then((module) => ({
-      default: module.InlineAlert,
-    })),
-  ),
-  "Content:@mittwald/flow-react-components/Content": lazy(() =>
-    import("@mittwald/flow-react-components/Content").then((module) => ({
-      default: module.Content,
-    })),
-  ),
-  "Label:@mittwald/flow-react-components/Label": lazy(() =>
-    import("@mittwald/flow-react-components/Label").then((module) => ({
-      default: module.Label,
-    })),
-  ),
-  "LabeledValue:@mittwald/flow-react-components/LabeledValue": lazy(() =>
-    import("@mittwald/flow-react-components/LabeledValue").then((module) => ({
-      default: module.LabeledValue,
-    })),
-  ),
-  "CopyToClipboardButton:@mittwald/flow-react-components/CopyToClipboardButton":
-    lazy(() =>
-      import("@mittwald/flow-react-components/CopyToClipboardButton").then(
-        (module) => ({ default: module.CopyToClipboardButton }),
-      ),
-    ),
-  "Link:@mittwald/flow-react-components/Link": lazy(() =>
-    import("@mittwald/flow-react-components/Link").then((module) => ({
-      default: module.Link,
-    })),
-  ),
-  "Navigation:@mittwald/flow-react-components/Navigation": lazy(() =>
-    import("@mittwald/flow-react-components/Navigation").then((module) => ({
-      default: module.Navigation,
-    })),
-  ),
-  "NavigationItem:@mittwald/flow-react-components/Navigation": I22,
-  "faUser:@fortawesome/free-regular-svg-icons/faUser": I23,
-  "faHardDrive:@fortawesome/free-regular-svg-icons/faHardDrive": I24,
-  "NumberField:@mittwald/flow-react-components/NumberField": lazy(() =>
-    import("@mittwald/flow-react-components/NumberField").then((module) => ({
-      default: module.NumberField,
-    })),
-  ),
-  "RadioGroup:@mittwald/flow-react-components/RadioGroup": lazy(() =>
-    import("@mittwald/flow-react-components/RadioGroup").then((module) => ({
-      default: module.RadioGroup,
-    })),
-  ),
-  "Radio:@mittwald/flow-react-components/RadioGroup": lazy(() =>
-    import("@mittwald/flow-react-components/RadioGroup").then((module) => ({
-      default: module.Radio,
-    })),
-  ),
-  "StatusIcon:@mittwald/flow-react-components/StatusIcon": lazy(() =>
-    import("@mittwald/flow-react-components/StatusIcon").then((module) => ({
-      default: module.StatusIcon,
-    })),
-  ),
-  "Switch:@mittwald/flow-react-components/Switch": lazy(() =>
-    import("@mittwald/flow-react-components/Switch").then((module) => ({
-      default: module.Switch,
-    })),
-  ),
-  "FieldDescription:@mittwald/flow-react-components/FieldDescription": lazy(
-    () =>
-      import("@mittwald/flow-react-components/FieldDescription").then(
-        (module) => ({ default: module.FieldDescription }),
-      ),
-  ),
-  "TextArea:@mittwald/flow-react-components/TextArea": lazy(() =>
-    import("@mittwald/flow-react-components/TextArea").then((module) => ({
-      default: module.TextArea,
-    })),
-  ),
-  "TextField:@mittwald/flow-react-components/TextField": lazy(() =>
-    import("@mittwald/flow-react-components/TextField").then((module) => ({
-      default: module.TextField,
-    })),
-  ),
-  "Tooltip:@mittwald/flow-react-components/Tooltip": lazy(() =>
-    import("@mittwald/flow-react-components/Tooltip").then((module) => ({
-      default: module.Tooltip,
-    })),
-  ),
-  "TooltipTrigger:@mittwald/flow-react-components/Tooltip": lazy(() =>
-    import("@mittwald/flow-react-components/Tooltip").then((module) => ({
-      default: module.TooltipTrigger,
-    })),
-  ),
-  "faSave:@fortawesome/free-regular-svg-icons/faSave": I35,
+  "ExampleSvg:@/content/01-getting-started/stylesheet/examples/components/ExampleSvg": lazy(() => import("@/content/01-getting-started/stylesheet/examples/components/ExampleSvg").then(module => ({ default: module.ExampleSvg } ))),
+"Avatar:@mittwald/flow-react-components/Avatar": lazy(() => import("@mittwald/flow-react-components/Avatar").then(module => ({ default: module.Avatar } ))),
+"Initials:@mittwald/flow-react-components/Initials": lazy(() => import("@mittwald/flow-react-components/Initials").then(module => ({ default: module.Initials } ))),
+"Image:@mittwald/flow-react-components/Image": lazy(() => import("@mittwald/flow-react-components/Image").then(module => ({ default: module.Image } ))),
+"Badge:@mittwald/flow-react-components/Badge": lazy(() => import("@mittwald/flow-react-components/Badge").then(module => ({ default: module.Badge } ))),
+"Text:@mittwald/flow-react-components/Text": lazy(() => import("@mittwald/flow-react-components/Text").then(module => ({ default: module.Text } ))),
+"IconEmail:@mittwald/flow-react-components/Icons": lazy(() => import("@mittwald/flow-react-components/Icons").then(module => ({ default: module.IconEmail } ))),
+"Button:@mittwald/flow-react-components/Button": lazy(() => import("@mittwald/flow-react-components/Button").then(module => ({ default: module.Button } ))),
+"IconHome:@mittwald/flow-react-components/Icons": lazy(() => import("@mittwald/flow-react-components/Icons").then(module => ({ default: module.IconHome } ))),
+"Checkbox:@mittwald/flow-react-components/Checkbox": lazy(() => import("@mittwald/flow-react-components/Checkbox").then(module => ({ default: module.Checkbox } ))),
+"Heading:@mittwald/flow-react-components/Heading": lazy(() => import("@mittwald/flow-react-components/Heading").then(module => ({ default: module.Heading } ))),
+"Icon:@mittwald/flow-react-components/Icon": lazy(() => import("@mittwald/flow-react-components/Icon").then(module => ({ default: module.Icon } ))),
+"IconStar:@tabler/icons-react": lazy(() => import("@tabler/icons-react").then(module => ({ default: module.IconStar } ))),
+"InlineAlert:@mittwald/flow-react-components/InlineAlert": lazy(() => import("@mittwald/flow-react-components/InlineAlert").then(module => ({ default: module.InlineAlert } ))),
+"Content:@mittwald/flow-react-components/Content": lazy(() => import("@mittwald/flow-react-components/Content").then(module => ({ default: module.Content } ))),
+"Label:@mittwald/flow-react-components/Label": lazy(() => import("@mittwald/flow-react-components/Label").then(module => ({ default: module.Label } ))),
+"LabeledValue:@mittwald/flow-react-components/LabeledValue": lazy(() => import("@mittwald/flow-react-components/LabeledValue").then(module => ({ default: module.LabeledValue } ))),
+"CopyToClipboardButton:@mittwald/flow-react-components/CopyToClipboardButton": lazy(() => import("@mittwald/flow-react-components/CopyToClipboardButton").then(module => ({ default: module.CopyToClipboardButton } ))),
+"LayoutCard:@mittwald/flow-react-components/LayoutCard": lazy(() => import("@mittwald/flow-react-components/LayoutCard").then(module => ({ default: module.LayoutCard } ))),
+"Link:@mittwald/flow-react-components/Link": lazy(() => import("@mittwald/flow-react-components/Link").then(module => ({ default: module.Link } ))),
+"Navigation:@mittwald/flow-react-components/Navigation": lazy(() => import("@mittwald/flow-react-components/Navigation").then(module => ({ default: module.Navigation } ))),
+"NavigationItem:@mittwald/flow-react-components/Navigation": I21,
+"IconCustomer:@mittwald/flow-react-components/Icons": lazy(() => import("@mittwald/flow-react-components/Icons").then(module => ({ default: module.IconCustomer } ))),
+"IconProject:@mittwald/flow-react-components/Icons": lazy(() => import("@mittwald/flow-react-components/Icons").then(module => ({ default: module.IconProject } ))),
+"IconServer:@mittwald/flow-react-components/Icons": lazy(() => import("@mittwald/flow-react-components/Icons").then(module => ({ default: module.IconServer } ))),
+"NumberField:@mittwald/flow-react-components/NumberField": lazy(() => import("@mittwald/flow-react-components/NumberField").then(module => ({ default: module.NumberField } ))),
+"RadioGroup:@mittwald/flow-react-components/RadioGroup": lazy(() => import("@mittwald/flow-react-components/RadioGroup").then(module => ({ default: module.RadioGroup } ))),
+"Radio:@mittwald/flow-react-components/RadioGroup": lazy(() => import("@mittwald/flow-react-components/RadioGroup").then(module => ({ default: module.Radio } ))),
+"Section:@mittwald/flow-react-components/Section": lazy(() => import("@mittwald/flow-react-components/Section").then(module => ({ default: module.Section } ))),
+"TextField:@mittwald/flow-react-components/TextField": lazy(() => import("@mittwald/flow-react-components/TextField").then(module => ({ default: module.TextField } ))),
+"StatusIcon:@mittwald/flow-react-components/StatusIcon": lazy(() => import("@mittwald/flow-react-components/StatusIcon").then(module => ({ default: module.StatusIcon } ))),
+"Switch:@mittwald/flow-react-components/Switch": lazy(() => import("@mittwald/flow-react-components/Switch").then(module => ({ default: module.Switch } ))),
+"FieldDescription:@mittwald/flow-react-components/FieldDescription": lazy(() => import("@mittwald/flow-react-components/FieldDescription").then(module => ({ default: module.FieldDescription } ))),
+"TextArea:@mittwald/flow-react-components/TextArea": lazy(() => import("@mittwald/flow-react-components/TextArea").then(module => ({ default: module.TextArea } ))),
+"Tooltip:@mittwald/flow-react-components/Tooltip": lazy(() => import("@mittwald/flow-react-components/Tooltip").then(module => ({ default: module.Tooltip } ))),
+"TooltipTrigger:@mittwald/flow-react-components/Tooltip": lazy(() => import("@mittwald/flow-react-components/Tooltip").then(module => ({ default: module.TooltipTrigger } ))),
+"IconCopy:@mittwald/flow-react-components/Icons": lazy(() => import("@mittwald/flow-react-components/Icons").then(module => ({ default: module.IconCopy } ))),
 };